C++ readability review for ericdingle.
Review URL: http://codereview.appspot.com/4210046

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@187 15e8cca8-e42c-11de-a347-f34a4f72eb7d

+++ syzygy/pdb/pdb_stream.cc
@@ -11,17 +11,18 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 #include "syzygy/pdb/pdb_stream.h"
 
 namespace pdb {
 
-PdbStream::PdbStream(size_t length) : length_(length), pos_(0) {
+PdbStream::PdbStream(int length) : length_(length), pos_(0) {
 }
 
 PdbStream::~PdbStream() {
 }
 
-bool PdbStream::Seek(size_t pos) {
+bool PdbStream::Seek(int pos) {
   if (pos > length_)
     return false;
